Gather local candidate IPs and generate a certificate for each
This commit fixes an issue where LocalAddr for a given candidate
does not correspond to the "external" IP that the peer sees,
upon establishing an ICE connection. As a result of that issue,
the local certificate was generated for the wrong IP, resulting
in certificate mismatch with what's expected by the peer.

Instead, now we gather all local candidate IPs, generate a cert
for each and send it to the remote peer. The peer then picks the
cert corresponding to RemoteAddr of its end of the ICE connection.
This is a bit chattier than if we generated a single certificate
that included all local IPs, but the resulting concurrency logic
would have to be more complicated.